requirements.txt
flask
flask_restful
flask-jwt-extended
Flask-SQLAlchemy
PyJWT
psycopg2-binary
python-dotenv